# docs/geo-rd-map-openapi.yaml
openapi: 3.0.3
info:
  title: GEO RD MAP API
  version: 2.6.1
  description: API geoespacial central de República Dominicana para integrar división territorial, selectores Provincia →
    Ciudad → Barrio, geocodificación, autocomplete, reverse geocoding, lugares y direcciones, GeoJSON/MVT, geocercas, routing
    y navegación. Las integraciones externas usan una API key privada por aplicación, permisos por scope, restricciones de
    origen y límites de consumo. Esta especificación documenta únicamente la API pública de integración; los endpoints /api/v1/admin
    forman parte del panel operativo y no son una API de consumo.
  contact:
    name: Administración de GEO RD MAP
servers:
- url: https://geo.ltd.do
  description: Producción
- url: http://localhost:8080
  description: Backend local
security:
- ApiKeyAuth: []
paths:
...
  /api/v1/geo/reverse:
    get:
      tags:
      - Datos geoespaciales
      summary: Reverse geocoding territorial usando el índice Overture local
      parameters:
      - in: query
        name: lat
        required: true
        schema:
          type: number
          format: double
      - in: query
        name: lng
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: Respuesta correcta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '400':
          $ref: '#/components/responses/BadRequest'
      operationId: reverseOvertureTerritory
      security:
      - ApiKeyAuth: []
      x-required-scope: geo:read
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key privada de la aplicación. Las claves se crean en Integraciones API del panel administrativo. Cada
        clave puede limitarse por scopes, orígenes permitidos y rate limit.
...
  responses:
    BadRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: API key ausente, inválida o inactiva.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: La API key no posee el scope requerido o el origen no está autorizado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Límite global o por API key excedido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
      - valid
      - message
      properties:
        valid:
          type: boolean
          example: false
        message:
          type: string
          example: API Key inválida o inactiva
    GenericSuccessResponse:
      type: object
      required:
      - valid
      - data
      properties:
        valid:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
        message:
          type: string
...